The neutron-fwaas package should require ipset
Currently, ipset is a requirement for openstack-neutron.
It is possible that the neutron-fwaas will be installed on a node that
does not have openstack-neutron on it.

Due to the fact that the neutron-fwaas relies on iptables driver,it also needs
ipset to operate and therefore it should require it during installation.

This issue was first raised by I3ec50a6046fee77e3faf72f7e756c7ae18ebaf2a
which aims to solve the same issue for other neutron agents:
linuxbridge, macvtap, openvswitch and sriov-nic.

Related rhbz: 1321870

Change-Id: I754af8b6d109170364187297c8f92d43f45c35df